+++ MapleLib/WzLib/WzProperties/WzLongProperty.cs
@@ -79,7 +79,7 @@ namespace MapleLib.WzLib.WzProperties
         /// <summary>
         /// The value of the property
         /// </summary>
-        public long Value { get { return val; } set { val = value; } }
+        public long Value { get { return val; } set { val = (long) value; } }
         /// <summary>
         /// Creates a blank WzCompressedIntProperty
         /// </summary>
